requirements.txt
Flask==2.0.3
line-bot-sdk==2.2.1
plotly==5.6.0
